+++ client/src/components/RelatedProducts/Cards/Comparison.jsx
@@ -46,6 +46,7 @@ function Comparison() {
           </tr>
         </thead>
         <tbody>
+
           {FeatureSet.map((item, key) => (
             <tr key={key}>
               <Xfeature>{item1Array.map((item1) => ((item1.feature === item.feature) ? (item1.value) ? `${item1.value}` : '✓' : ' '))}</Xfeature>
